requirements.txt
Fiona==1.8.0
geopandas==0.4.0
grequests==0.3.0
numpy==1.16.2
pandas==0.23.4
Rtree==0.8.3
Shapely==1.6.4.post2
xlrd==1.1.0
us==1.0.0